==> notebooks/01_model_fit.Rmd <==
---
title: "Calculate metrics and plot model fit"
author: "JM Delgado"
date: "`r Sys.Date()`"
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
```

```{r}
library(dplyr)
library(RPostgreSQL)
library(ggplot2)
library(sf)

dataset=readRDS('../data/dataset.rds')

```

# Plot and calculate model fit metric

```{r}


dataset %>%
    select(id_jrc, time_stamp, volume_rs, volume_val, ref_area, algorithm) %>%
    group_by(id_jrc,time_stamp) %>% filter(n()==2) %>%
    tidyr::pivot_wider(names_from=algorithm, values_from=volume_rs) %>%
    ungroup %>%
    yardstick::mae(volume_val,`canny edge + concave hull`)

dataset %>%
    select(id_jrc, time_stamp, volume_rs, volume_val, ref_area, algorithm) %>%
    group_by(id_jrc,time_stamp) %>% filter(n()==2) %>%
    tidyr::pivot_wider(names_from=algorithm, values_from=volume_rs) %>%
    ungroup %>%
    yardstick::mae(volume_val,`minimum error threshold`)


## combined
dataset %>%
    select(id_jrc, time_stamp, volume_rs, volume_val, ref_area, algorithm) %>%
    group_by(id_jrc,time_stamp) %>% filter(n()==2) %>%
    tidyr::pivot_wider(names_from=algorithm, values_from=volume_rs) %>%
    ungroup %>%
    mutate(combined=(`minimum error threshold` + `canny edge + concave hull`)/2) %>%
    yardstick::mae(volume_val,combined)



plt=ggplot(dataset)+
  geom_point(aes(x=volume_rs,y=volume_val,color=id_jrc,shape=algorithm)) #+
#  coord_trans(x="log2", y="log2")
ggsave('validation_plot.png',plt)
```

The scatter plot is not good. There is no skill in predicting volume from satellite radar images. This could be expected, since the HAV function is calculated with maximum area and maximum perimeter obtained from a remote sensing dataset. Maybe this could be improved, but most importantly, _can we at least have some skill in predicting approximate states of the system?_ For example, predicting if the reservoir is empty, half-full or full? This would already be a great piece of information.
