==> notebooks/2_query_postgis_and_rasterize.ipynb <==
## Connect and query

import sys
sys.path.append('/home/delgado/proj/buhayra_manuscript/buhayra')


from credentials import *

import psycopg2
import osgeo.ogr
import shapely
import shapely.wkt
import geopandas as gpd

connection = psycopg2.connect(database="watermasks",user=postgis_user, password=postgis_pass,host=postgis_host)
cursor = connection.cursor()



## Derive query
The syntax of the query should look something like:

```
SELECT `id`, `id_jrc`, `ingestion_time`, `area`
FROM (SELECT `id`, `id_jrc`, `ingestion_time`, `area`, MAX(`ingestion_time`) OVER (PARTITION BY `id_jrc`) AS `q01`
FROM `df`)
WHERE (`ingestion_time` = `q01`)
```

I am splitting the query into a sub query `sub_q` and a main query. ***In the end I used only the subquery and performed the final filter on geopandas***

year=2020
month=4
 
sub_q_select = " SELECT id, id_jrc, ingestion_time, area, ST_AsText(geom), MAX(ingestion_time) OVER (PARTITION BY id_jrc)"
sub_q_from = " FROM manuscript_threshold_geom"
sub_q_where = " WHERE area>0 AND to_char(ingestion_time,'YYYY')='" + "{:0>2d}".format(year) + "' AND to_char(ingestion_time,'MM')='"+ "{:0>2d}".format(month) +"'"


sub_q = sub_q_select + sub_q_from + sub_q_where

#q_select = " SELECT id, id_jrc, ingestion_time, area, ST_AsText(geom)"
#q_from = " FROM (" + sub_q + ") AS q01"
#q_where=" WHERE (ingestion_time = q01)"

#query=q_select + q_from + q_where + ' LIMIT 50;'



cursor.execute(sub_q)

cursor.description

rows_list=[]
for index,id_jrc,ingestion_time,area,geo,max_ingest in cursor:
    data={'id':index,'id_jrc':id_jrc,'ingestion_time':ingestion_time,'area':area,'geometry':shapely.wkt.loads(geo),'max':max_ingest}
    rows_list.append(data)
gdf=gpd.GeoDataFrame(rows_list,crs='epsg:4326').set_index(['id'])



### Filter only latest extents in the given month for each id_jrc

gdf_unique=gdf.loc[gdf['max']==gdf['ingestion_time']].copy()

gdf_unique

### Create list of geometries in shapely

gdf_unique["value"]=1
shapes = ((geom, value) for geom, value in zip(gdf_unique.geometry, gdf_unique.value))

### Create template raster

import rasterio
rst = rasterio.open('../data/benchmark.tif')
meta = rst.meta.copy()
meta.update(compress='lzw')


meta["transform"]

### Burn into raster

from rasterio import features
with rasterio.open('../data/burned_raster.tif', 'w+', **meta) as out:
    out_arr = out.read(1)
#     burned = features.rasterize(shapes=shapes, fill=0, out=out_arr, transform=out.transform)
    burned = features.rasterize(shapes=shapes, out_shape=(meta['height'],meta['width']),fill=0, transform=meta['transform'])
    out.write_band(1, burned)

### Alternative without opening file

burned1 = features.rasterize(shapes=shapes, out_shape=[meta['height'],meta['width']],fill=0, transform=meta['transform'])
# burned2 = features.rasterize(shapes=shapes, out_shape=[meta['height'],meta['width']],fill=0, transform=meta['transform'])


### Still can save raster to file later

with rasterio.open('../data/burned_raster.tif', 'w+', **meta) as out:
    out.write_band(1,burned1)
