==> notebooks/3_compose_xarray.ipynb <==
# Compose xarray from monthly rasters

import xarray as xr

xds = xr.open_rasterio("../data/burned_raster.tif").squeeze(drop=True)

xdsc=xr.concat((xds, xds, xds), dim='month').assign_coords(month=[5,6,7])

xdsc.sel(month=6)
